+++ hitable.hpp
@@ -6,7 +6,7 @@
 class material;
 
 struct hit_record {
-    float t;
+    double t;
     vec3 p;
     vec3 normal;
     material const * pMat;
@@ -16,8 +16,8 @@ class hitable {
     public:
         virtual bool hit(
             ray const & r,
-            float t_min,
-            float t_max,
+            double t_min,
+            double t_max,
             hit_record & rec
         ) const = 0;
 };
