+++ ray.hpp
@@ -8,7 +8,7 @@ class ray {
         ray( vec3 const & a, vec3 const & b) { A = a; B = b; }
         vec3 const & origin() const {return A; }
         vec3 const & direction() const { return B; }
-        vec3 point_at_parameter(float const t) const {
+        vec3 point_at_parameter(double const t) const {
              return A + t * B;
         }
